implemented logger in ProductController, also modified createProduct method in the same class to display a message if the product is already in DB, implemented basic authentication with SecurityConfig.class and role based access

+++ src/main/java/com/storemanagementapi/StoreManagementAPI/controller/ProductController.java
@@ -3,6 +3,8 @@ package com.storemanagementapi.StoreManagementAPI.controller;
 
 import com.storemanagementapi.StoreManagementAPI.datamodel.Product;
 import com.storemanagementapi.StoreManagementAPI.repository.ProductRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,7 +18,7 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/api")
 public class ProductController {
-
+    private static final Logger LOG = LoggerFactory.getLogger(ProductController.class);
     @Autowired
     ProductRepository productRepository;
 
@@ -28,8 +30,10 @@ public class ProductController {
                 productRepository.findAll().forEach(products :: add);
             }
             if (products.isEmpty()){
+                LOG.info("No Products in DB");
                 return new ResponseEntity<>(products, HttpStatus.OK);
             }
+            LOG.info("Products returned");
             return new ResponseEntity<>(products,HttpStatus.OK);
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -41,8 +45,10 @@ public class ProductController {
         Optional<Product> _product = productRepository.findById(id);
         try {
             if (_product.isPresent()){
+                LOG.info("Product found by id");
                 return new ResponseEntity<>(_product.get(),HttpStatus.OK);
             }else {
+                LOG.info("No such id in DB");
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
         }catch (Exception e){
@@ -55,8 +61,10 @@ public class ProductController {
         try {
             Optional<Product> _product = productRepository.findByName(name);
             if (_product.isEmpty()){
+                LOG.info("No product with provided name exists in DB");
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
+            LOG.info("Product found");
             return new ResponseEntity<>(_product.get(), HttpStatus.OK);
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -67,9 +75,15 @@ public class ProductController {
     @PostMapping("/products")
     public ResponseEntity<Product> createProduct(@RequestBody Product product){
         try {
-            Product _product = productRepository
-                    .save(new Product(product.getName(), product.getDescription(), product.getPrice(), true,product.getCount()));
-            return new ResponseEntity<>(_product, HttpStatus.CREATED);
+            if (!productRepository.existsByName(product.getName())) {
+                Product _product = productRepository
+                        .save(new Product(product.getName(), product.getDescription(), product.getPrice(), true, product.getCount()));
+                LOG.info("Product created");
+                return new ResponseEntity<>(_product, HttpStatus.CREATED);
+            }else{
+                LOG.warn("Product is already in DB");
+                return new ResponseEntity<>(HttpStatus.IM_USED);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -86,8 +100,10 @@ public class ProductController {
                _product.setPrice(product.getPrice());
                _product.setInStock(true);
                _product.setCount(product.getCount());
+               LOG.info("Product successfully updated");
                return new ResponseEntity<>(productRepository.save(_product), HttpStatus.ACCEPTED);
            }else {
+               LOG.info("Product doesn't exist in DB");
                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
            }
        }catch (Exception e){
@@ -98,8 +114,14 @@ public class ProductController {
     @DeleteMapping("/products/{id}")
     public ResponseEntity<HttpStatus> deleteProduct(@PathVariable("id") long id){
         try {
-            productRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (productRepository.existsById(id)) {
+                productRepository.deleteById(id);
+                LOG.info("Product successfully deleted");
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else {
+                LOG.warn("Product id isn't in DB");
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -109,6 +131,7 @@ public class ProductController {
     public ResponseEntity<HttpStatus> deleteAllProducts(){
         try {
             productRepository.deleteAll();
+            LOG.info("All products where deleted");
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
